Crate solana_cli_config

Loading and saving the Solana CLI configuration file.

The configuration file used by the Solana CLI includes information about the RPC node to connect to, the path to the user’s signing source, and more. Other software than the Solana CLI may wish to access the same configuration and signer.

The default path to the configuration file can be retrieved from CONFIG_FILE, which is a lazy_static of Option<String>, the value of which is

~/.config/solana/cli/config.yml

CONFIG_FILE will only be None if it is unable to identify the user’s home directory, which should not happen under typical OS environments.

The CLI configuration is defined by the Config struct, and its value is loaded with Config::load and saved with Config::save.

Two important fields of Config are

json_rpc_url, the URL to pass to solana_rpc_client::rpc_client::RpcClient.

keypair_path, a signing source, which may be a keypair file, but may also represent several other types of signers, as described in the documentation for solana_clap_utils::keypair::signer_from_path.

Examples

Loading and saving the configuration. Note that this uses the anyhow crate for error handling.

use anyhow::anyhow;
use solana_cli_config::{CONFIG_FILE, Config};

let config_file = solana_cli_config::CONFIG_FILE.as_ref()
    .ok_or_else(|| anyhow!("unable to get config file path"))?;
let mut cli_config = Config::load(&config_file)?;
// Set the RPC URL to devnet
cli_config.json_rpc_url = "https://api.devnet.solana.com".to_string();
cli_config.save(&config_file)?;

Modules

config 🔒

config_input 🔒

Structs

CONFIG_FILE

The default path to the CLI configuration file.

Config

The Solana CLI configuration.

ConfigInput

Enums

SettingType

Functions

load_config_file

Load a value from a file in YAML format.

save_config_file

Save a value to a file in YAML format.