Struct solana_cli_config::config::CONFIG_FILE

pub struct CONFIG_FILE {
    __private_field: (),
}

The default path to the CLI configuration file.

This is a lazy_static of Option<String>, the value of which is

~/.config/solana/cli/config.yml

It will only be None if it is unable to identify the userโs home directory, which should not happen under typical OS environments.

Fields
__private_field: ()